requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic-settings>=2.2
cryptography>=42.0
bcrypt>=4.1
itsdangerous>=2.1

pytest>=8.0
httpx>=0.27
